+++ IPX800cs/Parsers/v4/M2M/IPX800V4GetGetAnalogInputM2MResponseParser.cs
@@ -2,7 +2,7 @@
 
 internal class IPX800V4GetGetAnalogInputM2MResponseParser : ResponseParserBase, IGetAnalogInputResponseParser
 {
-    public int ParseResponse(string ipxResponse, int inputNumber)
+    public double ParseResponse(string ipxResponse, int inputNumber)
     {
         return ParseValue(ipxResponse, inputNumber);
     }
